include/flags.yml: Build rust binaries with frame pointers
For x86_64 and aarch64.

rustc is currently following the same behavior as gcc on linux,
which is not ideal as we end up without frame pointers for x86_64
binaries.

In order to enalbe frame-pointers we need to define and export into
the build environment RUSTFLAGS much like we do with CFLAGS.

One omission is that currently we can't don't export flatpak-builder
defaults.json, but we can do so once flatpak-builder has support for
RUSTFLAGS too.

Close freedesktop-sdk/freedesktop-sdk#1552